requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-standardized-errors==0.12.5
drf-yasg==1.21.5
future==0.18.3
geocoder==1.38.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
nepali-datetime==1.0.7
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
ratelim==0.1.6
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==2.0.2
user-agents==2.2.0
